rename KVStorage to KVStore
